wip: adding tests for sqs
